Rename tests.rs to preinstalled_contracts.rs
